add more auctions and bids. more controle over currency, more bussiness rules

+++ src/main/java/com/subastas/virtual/dto/bid/http/BidRequest.java
@@ -11,5 +11,5 @@ public class BidRequest {
   private float bid;
 
   @JsonProperty("payment_method")
-  private Long paymentMethod;
+  private Integer paymentMethod;
 }
